src/main/java/com/demo/todo/list/app/security/exception/handling/TodoListAuthenticationEntryPoint.java
package com.demo.todo.list.app.security.exception.handling;

import com.demo.todo.list.app.model.response.ApiErrorResponse;
import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.AuthenticationEntryPoint;

import java.io.IOException;
import java.util.Date;

@RequiredArgsConstructor
@Slf4j
public class TodoListAuthenticationEntryPoint implements AuthenticationEntryPoint {

    private final ObjectMapper objectMapper;

    @Override
    public void commence(HttpServletRequest request,
                         HttpServletResponse response,
                         AuthenticationException authException) throws IOException {
        log.error("Authentication error: {}", authException.getMessage());
        response.setStatus(HttpStatus.UNAUTHORIZED.value());
        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
        response.setHeader("WWW-Authenticate", "Bearer");
        ApiErrorResponse apiErrorResponse = ApiErrorResponse.builder()
                .url(request.getRequestURI())
                .createdDate(new Date())
                .build();
        String result = objectMapper.writeValueAsString(apiErrorResponse);
        response.getWriter().write(result);
    }
}

